Make sure users can only delete their own profile

+++ server/src/controllers/user-controller.ts
@@ -1,6 +1,8 @@
-import { NextFunction, Request, Response } from "express";
-import { IUserService } from "../interfaces/user-service.js";
-import { catchAsync } from "../utils/catchAsync.js";
+import { NextFunction, Request, Response } from 'express';
+import { isValidObjectId } from 'mongoose';
+import { UserRequestWithId } from '../interfaces/requests.js';
+import { IUserService } from '../interfaces/user-service.js';
+import { catchAsync } from '../utils/catch-async.js';
 
 export class UserController {
   constructor(private userService: IUserService) {}
@@ -15,8 +17,8 @@ export class UserController {
       });
 
       res.status(201).json({
-        status: "success",
-        message: "User registered successfully!",
+        status: 'success',
+        message: 'User registered successfully!',
         _id: user._id,
         username,
         tokens,
@@ -28,11 +30,14 @@ export class UserController {
     async (req: Request, res: Response, next: NextFunction) => {
       const { username, password } = req.body;
 
-      const { user, tokens } = await this.userService.login({ username, password });
+      const { user, tokens } = await this.userService.login({
+        username,
+        password,
+      });
 
       res.status(200).json({
-        status: "success",
-        message: "Logged in successfully!",
+        status: 'success',
+        message: 'Logged in successfully!',
         _id: user._id,
         username,
         tokens,
@@ -40,21 +45,27 @@ export class UserController {
     },
   );
 
-  delete = catchAsync(
-    async (req: Request, res: Response, next: NextFunction) => {
-      const { id } = req.params;
-      const idString = id.toString();
+  delete = catchAsync(async (req: UserRequestWithId, res: Response) => {
+    const { id } = req.params;
+    const idString = id.toString();
 
-      const deleted =  await this.userService.delete(idString);
+    if (id == null || !isValidObjectId(id)) {
+      throw new Error('Invalid ID');
+    }
 
-      if (!deleted) {
-        throw new Error('User could not be deleted.');
-      }
+    if (req.user?.id != id) {
+      throw new Error('User not found.');
+    }
 
-      res.status(200).json({
-        status: "success",
-        message: "User deleted successfully!",
-      });
-    },
-  );
+    const deleted = await this.userService.delete(idString);
+
+    if (!deleted) {
+      throw new Error('User could not be deleted.');
+    }
+
+    res.status(200).json({
+      status: 'success',
+      message: 'User deleted successfully!',
+    });
+  });
 }
